requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
Flask-Limiter==3.8.0
Flask-Caching==2.3.0
SQLAlchemy==2.0.38
alembic==1.15.1
psycopg2-binary==2.9.10
Werkzeug==3.1.3
Jinja2==3.1.4
openai==1.63.0
gunicorn==23.0.0
itsdangerous==2.2.0
python-dotenv==1.0.1
